Refactored internals directory and some other functions

+++ tests/test_setup.py
@@ -2,11 +2,11 @@ import pandas as pd
 from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import LabelEncoder
 
-from pycarrot.modelling.internals import Setup
+from pycarrot.modelling._internals import Setup
 
 setup = Setup(
     X_train=pd.DataFrame(),
-    y_clf_train=pd.Series(),
+    y_clf_train=pd.Series(dtype="int"),
     y_clf_encoder=LabelEncoder(),
     prep_pipe=Pipeline(steps=[]),
 )
